+++ src/test/java/org/eclipse/tractusx/valueaddedservice/rest/DashBoardResourceIntegrationTest.java
@@ -19,17 +19,23 @@
 ********************************************************************************/
 package org.eclipse.tractusx.valueaddedservice.rest;
 
+import com.fasterxml.jackson.core.JsonProcessingException;
+import com.fasterxml.jackson.core.type.TypeReference;
 import com.fasterxml.jackson.databind.ObjectMapper;
 import lombok.extern.slf4j.Slf4j;
 import org.eclipse.tractusx.valueaddedservice.ValueAddedServiceApplication;
+import org.eclipse.tractusx.valueaddedservice.config.ApplicationVariables;
+import org.eclipse.tractusx.valueaddedservice.dto.AuthPropertiesDTO;
 import org.eclipse.tractusx.valueaddedservice.dto.DashBoardTableDTO;
 import org.eclipse.tractusx.valueaddedservice.dto.DashBoardWorldMapDTO;
 import org.eclipse.tractusx.valueaddedservice.dto.RatingDTO;
 import org.eclipse.tractusx.valueaddedservice.utils.PostgreSQLContextInitializer;
+import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.Test;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.beans.factory.annotation.Value;
 import org.springframework.boot.test.context.SpringBootTest;
+import org.springframework.boot.test.mock.mockito.MockBean;
 import org.springframework.boot.test.web.client.TestRestTemplate;
 import org.springframework.core.ParameterizedTypeReference;
 import org.springframework.core.io.Resource;
@@ -48,9 +54,10 @@ import java.util.Map;
 
 import static org.junit.jupiter.api.Assertions.assertEquals;
 import static org.junit.jupiter.api.Assertions.assertNotEquals;
+import static org.mockito.Mockito.when;
 
 @Slf4j
-@SpringBootTest(webEnvironment = SpringBootTest.WebEnvironment.DEFINED_PORT,classes = ValueAddedServiceApplication.class)
+@SpringBootTest(webEnvironment = SpringBootTest.WebEnvironment.RANDOM_PORT,classes = ValueAddedServiceApplication.class)
 @ContextConfiguration(initializers = PostgreSQLContextInitializer.class)
 class DashBoardResourceIntegrationTest {
 
@@ -63,6 +70,25 @@ class DashBoardResourceIntegrationTest {
 
     @Autowired
     ObjectMapper objectMapper;
+
+    @MockBean
+    ApplicationVariables applicationVariables;
+
+    @BeforeEach
+    public void setUp() throws JsonProcessingException {
+        AuthPropertiesDTO authPropertiesDTO = new AuthPropertiesDTO();
+        authPropertiesDTO.setCompanyName("TestCompany");
+        authPropertiesDTO.setEmail("test@email.com");
+        authPropertiesDTO.setName("TestName");
+
+        ObjectMapper mapper = new ObjectMapper();
+        String json = "{ \"Cl16-CX-CRisk\": { \"roles\": [ \"User\", \"Company Admin\", \"read_suppliers\", \"read_customers\" ] } }";
+        Map<String, Object> map = mapper.readValue(json, new TypeReference<Map<String, Object>>(){});
+        authPropertiesDTO.setResourceAccess(map);
+
+        when(applicationVariables.getAuthPropertiesDTO()).thenReturn(authPropertiesDTO);
+    }
+
 
 
     private Map<String,Object> getMap() throws IOException {